Custom dashboard

+++ routes/web.php
@@ -21,13 +21,13 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth'])->name('dashboard.')->group(function(){
     Route::view('/', 'dashboard.index')->name('index');
     Route::prefix('users')->name('users.')->group(function () {
-        Route::get('/', [UserController::class, 'index'])->name('index');
-        Route::get('/create', [UserController::class, 'create'])->name('create');
+        Route::get('/', [UserController::class, 'index'])->name('index')->middleware('permission:view admin dashboard');
+        Route::get('/create', [UserController::class, 'create'])->name('create')->middleware('permission:view admin dashboard');
         Route::post('/', [UserController::class, 'store'])->name('store');
         Route::get('/{user:username}', [UserController::class, 'show'])->name('show');
         Route::get('/{user:username}/edit', [UserController::class, 'edit'])->name('edit');
         Route::match(['put', 'patch'],'/{user:username}', [UserController::class, 'update'])->name('update');
-        Route::delete('/{user:username}', [UserController::class, 'destroy'])->name('destroy');
+        Route::delete('/{user:username}', [UserController::class, 'destroy'])->name('destroy')->middleware('permission:delete user profile');;
     });
 });
